Refactor conversion logic and relocate implementation
Revised conversion logic by replacing direct comparisons with helper functions for improved readability and accuracy. Moved `ConversionTrait` and `ConversionHelper` implementations to a new namespace for better organization and modularity. Ensured the conversion mechanism maintains its functionality while enhancing code maintainability.
