+++ platform/include/roughpy/generics/conversion_trait.h
@@ -6,6 +6,8 @@
 #define ROUGHPY_GENERICS_FROM_TRAIT_H
 
 
+#include "roughpy/core/check.h"
+#include "roughpy/core/debug_assertion.h"
 #include "roughpy/core/macros.h"
 
 #include "roughpy/platform/roughpy_platform_export.h"
@@ -113,9 +115,95 @@ public:
     RPY_NO_DISCARD Value convert(ConstRef src, bool exact = true) const;
 };
 
+namespace conv {
+/**
+ * @brief Represents the result of a type conversion operation.
+ *
+ * ConversionResult is an enumeration used to indicate the outcome of
+ * a conversion process, which can either be successful, inexact, or failed.
+ */
+enum class ConversionResult
+{
+    Success,
+    Inexact,
+    Failed
+};
 
 
+template <typename From, typename To, typename=void>
+struct ConversionHelper
+{
+    using from_ptr = const From*;
+    using to_ptr = To*;
 
+    static constexpr bool is_possible = false;
+
+    // Set to true if the conversion is guaranteed to be exact
+    static constexpr bool is_always_exact = false;
+
+    // The conversion implementation. Should only check for an inexact
+    // conversion if the ensure_exact flag is set (and if the is_always_exact is
+    // false).
+    static ConversionResult convert(to_ptr dst, from_ptr src, bool ensure_exact);
+
+};
+
+
+/**
+ * @brief Specialized implementation of the ConversionTrait for specific types.
+ *
+ * The ConversionTraitImpl class provides concrete functionality for type
+ * conversion between specified source and destination types. It includes
+ * methods for checking if the conversion is exact and performing the conversion
+ * without safety checks. It ensures that the conversion between the specified
+ * types is possible at compile time.
+ *
+ * @tparam From The source type for the conversion.
+ * @tparam To The destination type for the conversion.
+ */
+template <typename From, typename To, typename SFINEA=void>
+class ConversionTraitImpl : public ConversionTrait
+{
+    using helper = ConversionHelper<From, To>;
+    static_assert(helper::is_possible, "From must be convertible to To");
+
+public:
+    ConversionTraitImpl(TypePtr from_type, TypePtr to_type)
+        : ConversionTrait(std::move(from_type), std::move(to_type)) {}
+
+    bool is_exact() const noexcept override;
+
+    void unsafe_convert(void* dst, const void* src, bool exact) const override;
+};
+
+template <typename From, typename To, typename SFINEA>
+bool ConversionTraitImpl<From, To, SFINEA>::is_exact() const noexcept
+{
+    return helper::is_always_exact;
+}
+
+template <typename From, typename To, typename SFINEA>
+void ConversionTraitImpl<From, To, SFINEA>::unsafe_convert(void* dst,
+    const void* src,
+    bool exact) const
+{
+    RPY_DBG_ASSERT_NE(dst, nullptr);
+    RPY_DBG_ASSERT_NE(src, nullptr);
+
+    auto* dst_ptr = static_cast<typename helper::to_ptr>(dst);
+    auto* src_ptr = static_cast<typename helper::from_ptr>(src);
+
+    auto result = helper::convert(dst_ptr, src_ptr, exact);
+    if (result == ConversionResult::Failed) {
+        throw std::runtime_error("conversion failed");
+    }
+    if (exact && ConversionResult::Inexact == result) {
+        RPY_THROW(std::runtime_error,
+            "conversion was required to be precise but was not");
+    }
+}
+
+}
 
 }// namespace rpy::generics
